A problem about `=/=`.
faster-minikanren only attaches =/= information to some of the variables involved in the disequality (https://github.com/michaelballantyne/faster-miniKanren#disequality). That works when we are sure we'll revisit the constraint if one of  those variables is assigned a value. But with SMT we don't know when an  SMT variable's domain is constrained to a single value.