Fix the previous problem of removing duplicate declares.
The old version removed duplicate declares in `z/reify-SM`. This is buggy, because when calling `z/reify-SM`, the current branch's state-M has not been updated. If we remove the duplicate declarein `z/reify-SM`, the current branch's state-M will lose the declare. PS. `decls` is global, there might be the same declare which was already added to `decls` by the other branch.

This bug will cause a crash in SMT solver (e.g. unknown constant _v62) with some special `gc-assumption-threshold`. The root cause is that after `z/gc!`, we will call `replay-if-needed`, it will resend all the commands in the current branch's state-M to SMT solver, but it might have lost some declare as I mentioned before.

The solution is making `z/reify-SM` only deal with reification and moving the responsibilities of removing duplicate declares to `z/local`.